requirements.txt
pydantic
pydantic[dotenv]
celery
celery[librabbitmq]
celery[redis]
kombu
fastapi==0.74.0
sqlalchemy
loguru==0.6.0
psycopg2-binary==2.9.3
redis==4.1.4
uvicorn==0.17.5
sentry-sdk==0.10.2
requests
pytest
alembic
strawberry-graphql[fastapi]
uvicorn[standard]
strawberry-graphql[debug-server]